# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dental Clinic Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List all patients
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by name, email, or phone
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create a new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get a patient by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Not found
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Not found
              
    delete:
      operationId: deletePatient
      tags: [patients]
      summary: Delete a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /patients/{id}/appointments:
    get:
      operationId: listPatientAppointments
      tags: [patients]
      summary: List appointments for a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
  /patients/{id}/bills:
    get:
      operationId: listPatientBills
      tags: [patients]
      summary: List bills for a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of bills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bill"

  # ── Appointments ──────────────────────────────────────────────────────────
...
  /inventory:    
    get:
      operationId: listInventoryItems
      tags: [inventory]
      summary: List inventory items
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"

    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Create an inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemInput"
      responses:
        "201":
          description: created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
...
  /inventory/{id}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get inventory item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory item found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
        "404":
          description: Not found

    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update an inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
        "404":
          description: Not found

    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete an inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

        
  
  # ── Employees ─────────────────────────────────────────────
...
components:
  schemas:
...
    Patient:
      type: object
      required: [id, firstName, lastName, email, phone, dateOfBirth, createdAt, updatedAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
          format: date
        address:
          type: ["string", "null"]
        insuranceProvider:
          type: ["string", "null"]
        insuranceNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PatientInput:
      type: object
      required: [firstName, lastName, email, phone, dateOfBirth]
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
          format: date
        address:
          type: string
        insuranceProvider:
          type: string
        insuranceNumber:
          type: string
        notes:
          type: string
    PatientUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
          format: date
        address:
          type: ["string", "null"]
        insuranceProvider:
          type: ["string", "null"]
        insuranceNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # ── Appointment ───────────────────────────────────────────────────────────
    Appointment:
      type: object
      required: [id, patientId, patientName, dentistName, appointmentDate, appointmentTime, durationMinutes, type, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: string
        dentistName:
          type: string
        appointmentDate:
          type: string
          format: date
        appointmentTime:
          type: string
          description: "HH:MM format"
        durationMinutes:
          type: integer
        type:
          type: string
          enum: [checkup, cleaning, filling, extraction, root_canal, crown, whitening, orthodontics, xray, consultation, other]
        status:
          type: string
          enum: [scheduled, confirmed, completed, cancelled, no_show]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Bill:
      type: object
      required: [id, patientId, patientName, totalAmount, amountPaid, status, dueDate, items, createdAt, updatedAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: string
        appointmentId:
          type: ["integer", "null"]
        totalAmount:
          type: number
        amountPaid:
          type: number
        status:
          type: string
          enum: [pending, partial, paid, overdue]
        dueDate:
          type: string
          format: date
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/BillLineItem"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BillLineItem:
      type: object
      required: [description, quantity, unitPrice]
      properties:
        description:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
...
    InventoryItem:
      type: object
      required: [id, name, category, quantity, unit, reorderLevel, costPerUnit, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        quantity:
          type: integer
        unit:
          type: string
        reorderLevel:
          type: integer
        costPerUnit:
          type: number
        supplier:
          type: ["string", "null"]
        lastRestockedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    InventoryItemInput:
      type: object
      required: [name, category, quantity, unit, reorderLevel, costPerUnit]
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
        quantity:
          type: integer
        unit:
          type: string
        reorderLevel:
          type: integer
        costPerUnit:
          type: number
        supplier:
          type: string
    InventoryItemUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        quantity:
          type: integer
        unit:
          type: string
        reorderLevel:
          type: integer
        costPerUnit:
          type: number
        supplier:
          type: ["string", "null"]
        lastRestockedAt:
          type: ["string", "null"]
          format: date-time

    # ── Employee ────────────────────────────────────────────────────────────
...