# lib/api-zod/src/generated/types/inventoryItemUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Dental Clinic Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryItemUpdate {
  name?: string;
  category?: string;
  quantity?: number;
  unit?: string;
  reorderLevel?: number;
  costPerUnit?: number;
  /** @nullable */
  supplier?: string | null;
  /** @nullable */
  lastRestockedAt?: Date | null;
}