# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Dental Clinic Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Patient {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  dateOfBirth: Date;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  insuranceProvider?: string | null;
  /** @nullable */
  insuranceNumber?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}